feat(*):完成使用minio进行文件的上传接口

+++ src/main/java/top/chopper/controller/SysDishController.java
@@ -41,7 +41,7 @@ public class SysDishController {
         LambdaQueryWrapper<SysDish> queryWrapper = new LambdaQueryWrapper<>();
         Page<SysDish> page = new Page<>(queryPageDto.getPage(),queryPageDto.getLimit());
         queryWrapper.like(queryPageDto.queryConditionIsExists("name"),SysDish::getName,queryPageDto.getQueryConditionValue("name"))
-                .eq(queryPageDto.queryConditionIsExists("type"),SysDish::getTypeId,(Integer)queryPageDto.getQueryConditionValue("type")) ;
+                .eq(queryPageDto.queryConditionIsExists("type"),SysDish::getTypeId,queryPageDto.getQueryConditionValue("type")) ;
         return R.SUCCESS( service.page(page,queryWrapper));
     }
